% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate4R-package.R
\name{check_installed_versions}
\alias{check_installed_versions}
\title{Check installed versions against DESCRIPTION}
\usage{
check_installed_versions()
}
\description{
Check installed versions against DESCRIPTION
}
