% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotes_check.R
\name{climate4R_check_remotes_consistency}
\alias{climate4R_check_remotes_consistency}
\title{Climate4R Remotes Consistency Check}
\usage{
climate4R_check_remotes_consistency()
}
\value{
Invisibly returns TRUE if all checks pass; otherwise, issues warnings.
}
\description{
Verifies that the versions declared in Imports match the versions (tags) specified in Remotes.
}
