% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{climate4R_report}
\alias{climate4R_report}
\title{Climate4R Full Diagnostic Report}
\usage{
climate4R_report()
}
\value{
Invisibly returns TRUE if no major inconsistencies are found; otherwise FALSE.
}
\description{
Generates a full report checking version compatibility, remotes consistency,
and listing installed component versions for Climate4R.
}
