% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_info.R
\name{climate4R_session_info}
\alias{climate4R_session_info}
\title{Climate4R Session Information}
\usage{
climate4R_session_info()
}
\value{
Invisibly returns a named vector of installed versions.
}
\description{
Lists the versions of the core Climate4R component packages that are required
and those currently installed. This function reads the required versions
dynamically from the DESCRIPTION file.
}
