% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate4R-package.R
\docType{package}
\name{climate4R}
\alias{climate4R}
\alias{climate4R-package}
\title{Climate4R: A Framework for Climate Data Management, Analysis, and Downscaling}
\description{
Climate4R provides a modular framework for managing, processing, analyzing, and visualizing climate data,
particularly focusing on datasets in NetCDF format. It integrates multiple specialized R packages,
offering a cohesive, reproducible, and scalable environment for climate data science and statistical downscaling.
}
\author{
\strong{Maintainer}: Santander Meteorology Group \email{meteo@ifca.es} (Institutional author)

}
