% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate4R-package.R
\name{climate4R_check_versions}
\alias{climate4R_check_versions}
\title{Check installed versions against DESCRIPTION}
\usage{
climate4R_check_versions()
}
\description{
Check installed versions against DESCRIPTION
}
